+++ common/api/decoupage_administratif.py
@@ -54,11 +54,31 @@ def map_epcis_code_name():
     Input: a list of epcis. {"nom":"CC Faucigny - Glières","code":"200000172","codesDepartements":["74"],"codesRegions":["84"],"population":28363}
     Output: a dict with the epci's code as pivot. {"200000172": "CC Faucigny - Glières"}
     """
-    epci_names = {}
+    epcis_names = {}
     try:
         epcis = fetch_epcis()
         for epci in epcis:
-            epci_names[epci["code"]] = epci["nom"]
+            epcis_names[epci["code"]] = epci["nom"]
     except requests.exceptions.HTTPError as e:
         logger.info(e)
-    return epci_names
+    return epcis_names
+
+
+def fetch_commune_detail(code_insee_commune, commune_details, geo_detail_type):
+    """
+    Provide EPCI code/ Department code/ Region code for a city, given the insee code of the city
+    """
+    if (
+        code_insee_commune
+        and code_insee_commune in commune_details.keys()
+        and geo_detail_type in commune_details[code_insee_commune].keys()
+    ):
+        return commune_details[code_insee_commune][geo_detail_type]
+
+
+def fetch_epci_name(code_insee_epci, epcis_names):
+    """
+    Provide EPCI code for an epci, given its insee code
+    """
+    if code_insee_epci and code_insee_epci in epcis_names.keys():
+        return epcis_names[code_insee_epci]
