+++ common/api/tests/test_decoupage_administratif.py
@@ -5,6 +5,8 @@ import requests_mock
 
 from common.api.decoupage_administratif import (
     DECOUPAGE_ADMINISTRATIF_API_URL,
+    fetch_commune_detail,
+    fetch_epci_name,
     map_communes_infos,
     map_epcis_code_name,
 )
@@ -51,6 +53,13 @@ class TestDecoupageAdministratifAPI(unittest.TestCase):
         self.assertEqual(communes_details["01002"]["city"], "L'Abergement-de-Varey")
         self.assertIsNone(communes_details["01002"]["epci"])  # Not all cities are part of an EPCI
 
+        self.assertEqual(fetch_commune_detail("01001", communes_details, "city"), "L'Abergement-Clémenciat")
+        self.assertEqual(fetch_commune_detail("01001", communes_details, "epci"), "200069193")
+        self.assertEqual(fetch_commune_detail("01002", communes_details, "city"), "L'Abergement-de-Varey")
+        self.assertEqual(fetch_commune_detail("01002", communes_details, "epci"), None)
+        self.assertEqual(fetch_commune_detail("01003", communes_details, "city"), None)
+        self.assertEqual(fetch_commune_detail("01003", communes_details, "epci"), None)
+
     def test_map_epcis_code_name(self, mock):
         mock.get(
             f"{DECOUPAGE_ADMINISTRATIF_API_URL}/epcis?fields=nom",
@@ -65,3 +74,7 @@ class TestDecoupageAdministratifAPI(unittest.TestCase):
         self.assertEqual(len(epci_names), 2)
         self.assertEqual(epci_names["200000172"], "CC Faucigny - Glières")
         self.assertEqual(epci_names["200000438"], "CC du Pays de Pontchâteau St-Gildas-des-Bois")
+
+        self.assertEqual(fetch_epci_name("200000172", epci_names), "CC Faucigny - Glières")
+        self.assertEqual(fetch_epci_name("200000438", epci_names), "CC du Pays de Pontchâteau St-Gildas-des-Bois")
+        self.assertEqual(fetch_epci_name("200000439", epci_names), None)
