+++ macantine/etl/open_data.py
@@ -9,7 +9,12 @@ from django.core.files.storage import default_storage
 from django.db.models import Q
 
 import macantine.etl.utils
-from common.api.decoupage_administratif import map_communes_infos, map_epcis_code_name
+from common.api.decoupage_administratif import (
+    fetch_commune_detail,
+    fetch_epci_name,
+    map_communes_infos,
+    map_epcis_code_name,
+)
 from data.models import Canteen
 from macantine.etl import etl
 from macantine.etl.etl import logger
@@ -34,18 +39,16 @@ class OPEN_DATA(etl.TRANSFORMER_LOADER):
         if "campagne_td" in self.dataset_name:
             # Get canteen geo data as most of TD don't have this info
             self.df["canteen_department"] = self.df["canteen_city_insee_code"].apply(
-                lambda x: macantine.etl.utils.fetch_commune_detail(x, communes_infos, "department")
+                lambda x: fetch_commune_detail(x, communes_infos, "department")
             )
             self.df["canteen_region"] = self.df["canteen_city_insee_code"].apply(
-                lambda x: macantine.etl.utils.fetch_commune_detail(x, communes_infos, "region")
+                lambda x: fetch_commune_detail(x, communes_infos, "region")
             )
             self.df["canteen_epci"] = self.df["canteen_epci"].apply(
-                lambda x: macantine.etl.utils.fetch_commune_detail(x, communes_infos, "epci")
+                lambda x: fetch_commune_detail(x, communes_infos, "epci")
             )
             epcis_names = map_epcis_code_name()
-            self.df["canteen_epci_lib"] = self.df["canteen_epci"].apply(
-                lambda x: macantine.etl.utils.fetch_epci_name(x, epcis_names)
-            )
+            self.df["canteen_epci_lib"] = self.df["canteen_epci"].apply(lambda x: fetch_epci_name(x, epcis_names))
 
         logger.info("Start filling geo_name")
         self.fill_geo_names(prefix)
